visualizer/requirements.txt
boto3>=1.26.0
flask>=2.2.0
gunicorn>=20.1.0
python-dotenv>=0.21.0
pandas>=1.5.0
